v0.1 Final Edit
There is no chance to the source code as of now. Everything works in
the same way as before, we must download the source code and source the
bashrc file that comes with it. In the next few days I will be porting
the code so that only python3 runs it and all the source code will be
available from pypi. This will be done in the development branch.