Reading Strings
We could still use `istringstream` but doing so will copy the string
that we wish to parse. To avoid this there is a method called
`pubsetbuf` but it seems that this may not be implemented in all
systems, meaning that this solution is not portable. David Godfrey
came up with a concise solution which I have included in
`stextInterface.h`. The adaptation is straight forward. If there is
every a problem we could always fall back on `istringstream`.