Default Paths
Since Python, MATLAB and compiled programs need be placed in a
directory where they can be found we will no longer place the output
in the current directory. Instead, they will be placed somewhere
close to the Python installation: `site.getuserbase()`. The paths
stated in the modifications are exported in the `excenturyrc` file.
There were some modifications to the python wrapper in this commit,
this is because there is a bug in the way we write the wrappers. The
fix will come next.