Python RuntimeError
When there are errors raised from within C++ there is always the
chance that the message provided will be deleted before Python gets
the message. This is the reason we have a static string in the
library for each function. This variable is to remain in the library
until Python has had the time to copy the message. This bug has now
been fix with these changes to `cpp_writer` in the Python language.