Main Script Edits
Minor changes to the main script. These are some changes that will be
required for Python 3. For the moment we need to create a few more
test cases to make sure that the Python conversions work. The
MANIFEST.in file was missing the header files so that we may package
the C++ libraries.