Empty Configuration Variable
Excentury has been tested in its own closed environment from its
moment of creation. Because of this there was no time to check for
other alternatives. The configuration module as it stands is great
for a single user. This however is not optimal for collaboration. We
can overwrite the values in the command line but it is likely that we
will forget how to overwrite them. More thought needs to be put into
the development of the configuration.