Configuration Interpolation
The configuration file parser now supports interpolation. That is, we
can refer to previously declared keys. For instance, in the current
modification to the `xcpp.config` file we can see that we use
`${path}`. This means that within that section we will obtain the
previous definition of path. This gets evaluated line by line.
Similarly to refer to a key in another section we use `${sec:key}` as
seen with `${xcpp:path}` in the last entry. Notice that the
configuration file also supports environmental variables expansion.