Manual Configuration
The `config` command used to allow us to modify the configuration
file. This idea was originally taken from how `git` allows us to
modify its configuration file. Since we have updated the way the
configuration file works we need to be able to modify this file
manually. In other words, excentury cannot rewrite this file for us.
When the command `config` is called with no arguments then it reads
and evaluates the file and prints all the values it holds. If you
need to see the original file without having to open it manually you
can use the option `--print`. As with `git` you can still obtain a
value from the configuration file if you wish as before.