Conditional Configuration
Some of the variables specified in the configuration may not work
correctly if you are working with collaborators or in another system.
For this reason we can now conditionally assign a value to a
variable. This is done by writing

    var = <valiftrue> IF[[ <condition> ]]

This will assign the value `<valiftrue>` to `var` only if the
condition is met. Otherwise var is left unchanged. The other type of
conditional assignment is

    var = <valiftrue> IF[[ <condition> ]]ELSE <valiffalse>

Similar to the previous one but this time if the condition is not met
we have the option of specifying the value the variable should take.

The only restrictions we have on `<condition>` is that it should be a
valid Python statement. Notice however, that the condition we write
will first be interpolated by the configuration parser. That is, if
we have `$HOME` written in there, it will first be expanded to its
value. After that, the statement will be evaluated by the `eval`
function.

NOTE: This could be potentially dangerous if we read a configuration
file from an untrusted source, but it can be helpful if used
correctly.

See the few examples in this commit to see the type of conditional
assignments we can make. The last example should give us an error
since Python thinks that jmlopez is an undefined variable. What we
actually meant to write was `'$USER' == 'jmlopez'`.