Python Adaptor
The inclusion of excentury was automatic. This however did not give
us a chance to include other libraries before excentury. To get
around this we were using `[[` and `]]`. The text written in between
these delimiters was called the `pre-xc` section. This section is
still in place but it will now be used to write comments that will
not appear in any cpp generated files.