CPP Adaptor
We must include excentury in the xcpp files. Similarly as in the last
commit, now the CPP adaptor uses this standard. One notable change we
must mention here is that the generated file does not need to include
the hook any longer. This is done by defining the macro `XC_CPP`
which tells the program the type of language the code will be
translated to. In this way excentury will be able to decide if it
needs to include the hook of not. This is similar as how we will
include Mathematica in the future.