MATLAB Adaptor
Due to a problem with some changes in Xcode in OS X the mex script
does not work: <http://stackoverflow.com/q/22367516/788553>.

Everything seems to work if we include `<algorithm>` before the
inclusion of `"mex.h"`. The inclusion of this file is now done as a
hook. The hook however needs to appear before all other inclusions.
If there is another problem we may have to modify this hook.

There is one problem with the execution of mex: it displays the
message "Configured with...". A quick fix to be able to proceed with
the compilation is to use the `-v` option but this will display a lot
of messages. A modification to the file has been done which locates
this message and continues the execution if only that message is
found.

Not all the functions have been tested yet although the few ones that
have been tested seem to work.