Excentury Projects
For the most part we will be working with projects and thus we need
to separate all the files so that they do not linger in the same
directory. By default excentury places those files in a special
directory. In a project however, many mistakes will be encountered.
When this is the case we would like to see the files generated by
excentury so that we may detect the source of the error.

The `init` command creates all the directories and a special bash
file so that the environmental paths point to the right directories.
Notice that this file should be modified to meet the projects
demands. For instance, a special directory where we place python
scripts could be created and we may append it to the `PYTHONPATH`
variable so that we do not mix the files generated by excentury and
those created by the team.

As a result of this idea the `install` command had to be modified.
Now we have a function to generate the standard `excenturyrc` file
which resides in `~/.excentury`. In the next commit we will create
another command which will make use of some of the functions in the
`install` command.