Select Projects
The select command will append the line

source path/to/xcpprc

to the `~/.excentury/excenturyrc` file. This allows us to modify the
bash_profile with specific project settings. To remove any of the
project settings we can select `0` to enter excenturys default
behavior.

When using the `init` command it will automatically call the `add`
command so that excentury can keep track of the project. We must
select the project after creating it, however.