Configuration Warning
If a project was selected previously we might forget that the
configuration was changed. This might give us problems if we try to
compile a temporary excentury file which is outside the project
scope. The configuration will let us know if it is reading from a
project, i.e., the environmental variable XCPP_CONFIG_PATH was set.
This should give us the hint as to why the file is not working. To
fix it we should deselect the project by calling `excentury select 0`
and restarting bash by either sourcing the bashrc file or by opening
a new terminal.

A mistake was fixed in the `init` command which created the wrong
path to the configuration file. As a result of this the `.xcpprc`
file in the tests directory was also fixed.