Includes and Libraries
An error was detected when trying to compile a projects xcpp file
from a different location other than its configuration file. This was
because the includes did not point to the correct directories.

This has now been fixed by appending the root. Note however, that the
paths need to be relative to the location of the configuration file.
If you need to include absolute paths then do so in the `xcpprc` file
which gives you more control. That is, if you are working in
different systems you may want to check in which system you are
working in. This could also be done with the configuration file but
overall we would like to discourage absolute paths in these files.