Unit Testing
A new xcpp file has been created so that we may test the
interoperability between C++ and the other languages. The tests are
very simple and they verify that excentury is working properly.

It should be noted that the test only work if we are using the
project. That is, the `xcpp.config` file will not help us much if we
do not source the `.xcpprc` file. Make sure to select the `xc-tests`
project before trying to run make.