Python3 Compatability
Making excentury work with Python3 might present a challenge since at
the moment I do not completely understand the difference between a
`string` and a `byte` in Python3. For now only small changes were
made since sphinx was downloaded with Python3 and in order to work it
needs to load excentury. Two changes have been made:

`long` is simply `int` in Python3 and the `cStringIO` library was
moved to `io`. To do the fix we have used the `six` library.

Finally, we will no longer track the changes to the documentation
html files. To do this we have used the `.gitignore` file.