[DOC]: What is Excentury?
Here I attempt to give a description of what Excentury will allow us
to do. An example is presented in MATLAB which is then written in
C++. This section is meant to try to get the reader to feel curious
about Excentury. It should entice the reader to think about how
excentury will attempt to serve as a bridge between C++ and the
scripting languages.