[DOC]: Simple Example
The getting started section has been simplified with a short example.
This should allow us to focus more on the excentury format instead of
trying to figure out what the function does. The installation makes
mention of issues in OS X. Finally, I have added the life `xc`. This
is so that we may include `<xc>` instead of
`<excentury/excentury.h>`. When this is done we also have the
namespace alias `xc`.