Armadillo Example
Configuration parser used to give an error if we tried to use a
variable which was not previously defined. A mistake in the macro
`XC_LOAD_TEMPLATED_TENSOR` was fixed. The reason we do not need to
define `f` is because the definition of this routine does not require
us to use `XC_DUMP` or `XC_LOAD` which do make use of `f`. The
armadillo example is now working in the tests.