Testing Armadillo Library
The incorporation of the matrix objects seem to work rather well with
the exception of a matrix of `float`s. The culprit was the way we
provide a sample of the matrix to excentury. By using `item(0)` we
are providing

> A copy of the specified element of the array as a suitable Python
> scalar.

This is problem since we want to keep the information of the matrix
intact. Please see the file `test_arma.py` to see how to correctly
call the example using floats instead of doubles.